# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HomeResponse {
  message: string;
}

export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  name: string;
  phone: string;
  balance: number;
  /** Write-only account password; never returned by the API. */
  password?: string;
}

export interface UserInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  /**
     * @minLength 8
     * @maxLength 72
     */
  password: string;
}

export interface LoginRequest {
  /** @minLength 1 */
  phone: string;
  /**
     * @minLength 1
     * @maxLength 72
     */
  password: string;
}

export interface LoginResponse {
  accessToken: string;
  refreshToken: string;
  user: User;
}

export interface RefreshRequest {
  /** @minLength 1 */
  refreshToken: string;
}

export interface ErrorResponse {
  error: string;
}