# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: users
    description: AartPay Tech user accounts and authentication
paths:
  /:
    get:
      operationId: getApiHome
      tags: [health]
      summary: API home
      description: Returns a welcome message for the AartPay Tech API
      responses:
        "200":
          description: API welcome message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeResponse"
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /register:
    post:
      operationId: registerUser
      tags: [users]
      summary: Register a user
      description: Creates a user with a starting balance of zero
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Registered user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid user data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A user with this phone number already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List registered users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Registered users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Log in with a phone number and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated user and token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid login request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid phone number or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /refresh:
    post:
      operationId: refresh
      tags: [users]
      summary: Rotate an access and refresh token pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Refreshed token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid refresh request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get the authenticated user's profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Authenticated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HomeResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        balance:
          type: number
        password:
          type: string
          writeOnly: true
          description: Write-only account password; never returned by the API.
      required:
        - id
        - name
        - phone
        - balance
    UserInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 8
          maxLength: 72
      required:
        - name
        - phone
        - password
    LoginRequest:
      type: object
      properties:
        phone:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
          maxLength: 72
      required:
        - phone
        - password
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - accessToken
        - refreshToken
        - user
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          minLength: 1
      required:
        - refreshToken
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error