# lib/api-zod/src/generated/types/loginRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LoginRequest {
  /** @minLength 1 */
  phone: string;
  /**
     * @minLength 1
     * @maxLength 72
     */
  password: string;
}