# lib/api-zod/src/generated/types/loginResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface LoginResponse {
  accessToken: string;
  refreshToken: string;
  user: User;
}