# lib/api-zod/src/generated/types/refreshRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RefreshRequest {
  /** @minLength 1 */
  refreshToken: string;
}